Fix for djang-1.9 tix:
Added two migrations, one to alter the FK field on UserSocialAuth to point it to 'ttusers.TTUser',
this should not actually be run for real and should be faked. I think it's because in our original
south_migrations there never was a swappable_dependency defined?

The other migration fixes the default field for python-social-auth's
inbuilt JSONField. It was previously returning as unicode/str which is incorrect.